+++ addons/sapian_demo_trader/tests/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import test_demo_trader_e2e
+from . import test_catalog_dependencies